+++ micro-infra-spring-base/src/main/java/com/ofg/infrastructure/hystrix/IsHystrixServletEnabled.java
@@ -1,0 +1,15 @@
+package com.ofg.infrastructure.hystrix;
+
+import com.ofg.infrastructure.config.PropertyAbsentOrEnabledCondition;
+
+/**
+ * Checks for the presence of the property to verify whether Hystrix is enabled or not
+ */
+class IsHystrixServletEnabled extends PropertyAbsentOrEnabledCondition {
+
+    public static final String HYSTRIX_SERVLET_ENABLED = "com.ofg.infra.microservice.hystrix.servlet";
+
+    public IsHystrixServletEnabled() {
+        super(HYSTRIX_SERVLET_ENABLED);
+    }
+}
